UPDATE ::: Additional checks for generating / updating the insights database
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@30884 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

+++ src/prerna/engine/impl/InsightsDatabaseUpdater3CacheableColumn.java
@@ -1,8 +1,10 @@
 package prerna.engine.impl;
 
+import java.sql.ResultSet;
 import java.sql.SQLException;
 
 import prerna.engine.impl.rdbms.RDBMSNativeEngine;
+import prerna.sablecc2.reactor.app.upload.UploadUtilities;
 
 @Deprecated
 public class InsightsDatabaseUpdater3CacheableColumn {
@@ -26,29 +28,56 @@ public class InsightsDatabaseUpdater3CacheableColumn {
 
 	
 	@Deprecated
-	public static void update(RDBMSNativeEngine insightsRdbms) {
-		// first let us drop the tables we do not require
-		{
-			for(String col : COLUMNS_TO_DELETE) {
-				String dropColQuery = "ALTER TABLE QUESTION_ID DROP COLUMN IF EXISTS " + col;
-				try {
-					insightsRdbms.removeData(dropColQuery);
-				} catch (SQLException e) {
-					e.printStackTrace();
-				}
-			}
-		}
+	public static void update(String appId, RDBMSNativeEngine insightsRdbms) {
 		
-		// let us add the new column with default value of true
-		{
+		boolean tableExists = false;
+		ResultSet rs = null;
+		try {
+			rs = insightsRdbms.getConnectionMetadata().getTables(null, null, "QUESTION_ID", null);
+			if (rs.next()) {
+				  tableExists = true;
+			}
+		} catch (SQLException e) {
+			e.printStackTrace();
+		} finally {
 			try {
-				insightsRdbms.insertData(UPDATE_QUERY);
-			} catch (SQLException e) {
+				if(rs != null) {
+					rs.close();
+				}
+			} catch(SQLException e) {
 				e.printStackTrace();
 			}
 		}
 		
-		insightsRdbms.commit();
+		if(!tableExists) {
+			// well, you already created the file
+			// need to run the queries to make this
+			UploadUtilities.runInsightCreateTableQueries(insightsRdbms);
+			UploadUtilities.addExploreInstanceInsight(appId, insightsRdbms);
+		} else {
+			// first let us drop the tables we do not require
+			{
+				for(String col : COLUMNS_TO_DELETE) {
+					String dropColQuery = "ALTER TABLE QUESTION_ID DROP COLUMN IF EXISTS " + col;
+					try {
+						insightsRdbms.removeData(dropColQuery);
+					} catch (SQLException e) {
+						e.printStackTrace();
+					}
+				}
+			}
+			
+			// let us add the new column with default value of true
+			{
+				try {
+					insightsRdbms.insertData(UPDATE_QUERY);
+				} catch (SQLException e) {
+					e.printStackTrace();
+				}
+			}
+			
+			insightsRdbms.commit();
+		}
 	}
 	
 }
